requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.0
email-validator==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1